Refactor du Security de Koala Rock et suppression de ce meme fichier
