+++ src/main/java/games/project/koala_rock/Metier/entite/AuthPlayer.java
@@ -1,7 +1,5 @@
 package games.project.koala_rock.Metier.entite; //Votre package ici.
-
-
-import games.project.koala_rock.Stockage.Security;
+import games.project.stockage.Security;
 
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
